requirements.txt
tensorflow
numpy
matplotlib
scipy
opencv-python
pandas
Pillow
ultralytics
